+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support of **sub-mesh-id** property was added into **CStaticMeshContainer** component.
 
-- The rendering of sub-meshes for static meshes was implemented.
+- The rendering of sub-meshes for static and skinned meshes was implemented.
 
 ### Changed